Move server routes to server folder
