+++ lib/server/index.js
@@ -5,7 +5,7 @@ import webpackDevMiddleware from 'webpack-dev-middleware'
 import webpackHotMiddleware from 'webpack-hot-middleware'
 import webpackConfig from '../../webpack/dev.webpack.config'
 
-import router from '../routes'
+import router from './routes'
 
 const server = Express()
 const port = process.env.PORT || 7002
