+++ lib/server/routes.js
@@ -1,6 +1,6 @@
 import express from 'express'
-import views from './views'
-import api from './api'
+import views from '../views'
+import api from '../api'
 
 const router = express.Router()
 export default router
